[bugfix] l:cljfmt_output locked on failure
From the Vim docs:

"After the redirection starts, if the variable is removed or locked or
the variable type is changed, then further command output messages will
cause errors."

I believe at the moment the fact that the redirection happens in a
try/catch block causes the redir to start, an error to happen in
fireplace, and then the redir is not ended so the variable is not
released. I'm not sure, but I think this patch should fix the problem.